requirements.txt
openai>=1.40.0
duckdb>=0.10.2
python-dotenv>=1.0.0
streamlit==1.39.0
requests>=2.20.0
altair==5.4.1
pandas==2.2.3
typing_extensions>=4.12.0